Corrected `labels` dtype in private test on `ArgKminLabels` directly

+++ sklearn/metrics/tests/test_pairwise_distances_reduction.py
@@ -669,7 +669,7 @@ def test_argkminlabels_factory_method_wrong_usages():
     metric = "euclidean"
 
     weights = "uniform"
-    labels = rng.randint(low=0, high=10, size=100)
+    labels = rng.randint(low=0, high=10, size=100, dtype=np.intp)
     unique_labels = np.unique(labels)
 
     msg = (
